fix: add filesize

+++ services/storage/models/storeModel.go
@@ -11,7 +11,7 @@ type Storage struct {
 	ID            uuid.UUID `gorm:"type:char(36);primary key" json:"id"`
 	Userid        uuid.UUID `gorm:"type:char(36)" json:"userid"`
 	FileName      string    `json:"file_name"`
-	FileSize      int       `json:"file_size"`
+	FileSize      int64     `json:"file_size"`
 	Mime_type     string    `json:"mime_type"`
 	EncryptedData []byte    `json:"file_path"`
 	UploadedAt    time.Time `json:"uploaded_at"`
